requirements.txt
audioread==2.1.9
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
docopt==0.6.2
idna==3.2
musicbrainzngs==0.7.1
mutagen==1.45.1
Pillow==8.3.2
pipreqs==0.4.10
pyacoustid==1.2.2
python-mpd2==3.0.5
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
yarg==0.1.9
discid==1.2.0
ffmpeg-python==0.2.0
redis==4.1.3
jsonpath-ng==1.5.3
aiofiles==0.6.0
celery==5.2.7
celery-progress
fastapi==0.79.0
flower
Jinja2==3.0.3
uvicorn[standard]
python-multipart
sqlalchemy
pytest==7.1.2
pyyaml
